--- packages/code-du-travail-frontend/src/dsfr/config/colors.ts
@@ -1,8 +1,0 @@
-export const colors = {
-  text: {
-    tint: "#000091",
-  },
-  background: {
-    lightBlue: "#e8edff",
-  },
-};
