+++ packages/code-du-travail-frontend/src/modules/config/defaultColorScheme.ts
