+++ packages/code-du-travail-frontend/src/modules/layout/NeedMoreInfo.tsx
@@ -1,14 +1,12 @@
 import { fr } from "@codegouvfr/react-dsfr";
 import { Button } from "@codegouvfr/react-dsfr/Button";
-import { colors } from "../config/colors";
 
 export const NeedMoreInfo = () => {
   return (
     <div style={styles.mainContainer}>
       <div className={fr.cx("fr-container")}>
         <div className={fr.cx("fr-grid-row")}>
-          <div className={fr.cx("fr-col-3")} />
-          <div className={fr.cx("fr-btns-group--center", "fr-col-6")}>
+          <div className={fr.cx("fr-grid-row--center")}>
             <h2 style={styles.title}>Besoin de plus d&apos;informations ?</h2>
             <p style={{ ...styles.paragraph, textAlign: "left" }}>
               Les services du ministère du Travail en région informent,
@@ -16,14 +14,13 @@ export const NeedMoreInfo = () => {
               privé sur leurs questions en droit du travail.
             </p>
             <Button
-              iconId="fr-icon-discuss-line"
+              iconId="fr-icon-discuss-fill"
               iconPosition="right"
               priority="tertiary"
             >
               Trouver les services près de chez moi
             </Button>
           </div>
-          <div className={fr.cx("fr-col-3")} />
         </div>
       </div>
     </div>
@@ -32,13 +29,13 @@ export const NeedMoreInfo = () => {
 
 const styles = {
   mainContainer: {
-    backgroundColor: `${colors.background.lightBlue}`,
+    // backgroundColor: `${colors.background.lightBlue}`,
     padding: `${fr.spacing("10v")} 0`,
   },
   paragraph: {
-    color: `${colors.text.tint}`,
+    // color: `${colors.text.tint}`,
   },
   title: {
-    color: `${colors.text.tint}`,
+    // color: `${colors.text.tint}`,
   },
 };
