+++ packages/code-du-travail-frontend/src/modules/layout/common.ts
