+++ packages/code-du-travail-frontend/src/modules/layout/header/__tests__/fetchSuggestResults.test.ts
